Successful pipeline run to make JSON data
