+++ gateway/sds_gateway/api_methods/__init__.py
@@ -1,6 +1,3 @@
-# Import cogs
-from .cog_pipelines import cleanup_temp_files_cog  # noqa: F401
-from .cog_pipelines import download_capture_files_cog  # noqa: F401
+# Import cogs for django-cog automatic discovery
 from .cog_pipelines import process_waterfall_data_cog  # noqa: F401
-from .cog_pipelines import store_processed_data_cog  # noqa: F401
-from .cog_pipelines import update_processing_status_cog  # noqa: F401
+from .cog_pipelines import setup_post_processing_cog  # noqa: F401
