+++ gateway/sds_gateway/api_methods/cog_pipelines.py
@@ -1,226 +1,18 @@
 """Django-cog pipeline configurations for post-processing."""
 
+import base64
+import datetime
+import os
+import tempfile
+from pathlib import Path
 from typing import Any
 
+import h5py
+import numpy as np
+from digital_rf import DigitalRFReader
+from django.conf import settings
 from django_cog import cog
 from loguru import logger
-
-
-@cog
-def process_waterfall_data_cog(
-    capture_uuid: str, max_slices: int = 100
-) -> dict[str, Any]:
-    """Process waterfall data for a capture.
-
-    Args:
-        capture_uuid: UUID of the capture to process
-        max_slices: Maximum number of slices to process (default: 100)
-
-    Returns:
-        Dict with status and processed data info
-    """
-    logger.info(f"Processing waterfall JSON data for capture {capture_uuid}")
-
-    try:
-        # Import models here to avoid Django app registry issues
-        from .models import Capture
-        from .models import ProcessingType
-        from .tasks import convert_drf_to_waterfall_json
-        from .tasks import reconstruct_drf_files
-        from .tasks import store_processed_data
-
-        capture = Capture.objects.get(uuid=capture_uuid, is_deleted=False)
-
-        # Create temporary directory for processing
-        import os
-        import tempfile
-        from pathlib import Path
-
-        temp_dir = tempfile.mkdtemp(prefix=f"waterfall_{capture_uuid}_")
-        temp_path = Path(temp_dir)
-
-        # Reconstruct the DigitalRF files for processing
-        capture_files = capture.files.filter(is_deleted=False)
-        reconstructed_path = reconstruct_drf_files(capture, capture_files, temp_path)
-
-        if not reconstructed_path:
-            return {
-                "status": "error",
-                "message": "Failed to reconstruct DigitalRF directory structure",
-            }
-
-        # Process the waterfall data in JSON format
-        waterfall_result = convert_drf_to_waterfall_json(
-            reconstructed_path,
-            capture.channel,
-            ProcessingType.Waterfall.value,
-            max_slices,
-        )
-
-        if waterfall_result["status"] != "success":
-            return waterfall_result
-
-        # Create a temporary JSON file
-        import json
-
-        with tempfile.NamedTemporaryFile(
-            mode="w", suffix=".json", delete=False
-        ) as temp_file:
-            json.dump(waterfall_result["json_data"], temp_file, indent=2)
-            temp_file_path = temp_file.name
-
-        # Store the JSON file
-        filename = f"waterfall_{capture_uuid}.json"
-        store_result = store_processed_data(
-            capture_uuid,
-            ProcessingType.Waterfall.value,
-            temp_file_path,
-            filename,
-            waterfall_result["metadata"],
-        )
-
-        # Clean up temporary file
-        os.unlink(temp_file_path)
-
-        logger.info(f"Completed waterfall processing for capture {capture_uuid}")
-        return {
-            "status": "success",
-            "capture_uuid": capture_uuid,
-            "message": "Waterfall JSON data processed and stored successfully",
-            "json_data": waterfall_result["json_data"],
-            "metadata": waterfall_result["metadata"],
-            "store_result": store_result,
-        }
-
-    except Exception as e:
-        logger.error(f"Waterfall processing failed for capture {capture_uuid}: {e}")
-        raise
-
-
-@cog
-def update_processing_status_cog(
-    capture_uuid: str, processing_type: str, status: str, step: str | None = None
-) -> dict[str, Any]:
-    """Update the processing status for a specific processing type.
-
-    Args:
-        capture_uuid: UUID of the capture
-        processing_type: Type of processing
-        status: Status to set (processing, completed, failed)
-        step: Optional step name for processing status
-
-    Returns:
-        Dict with status update info
-    """
-    try:
-        # Import models here to avoid Django app registry issues
-        from .models import Capture
-        from .models import PostProcessedData
-
-        capture = Capture.objects.get(uuid=capture_uuid)
-        processed_data = PostProcessedData.objects.filter(
-            capture=capture,
-            processing_type=processing_type,
-        ).first()
-
-        if processed_data:
-            if status == "processing":
-                processed_data.mark_processing_started(
-                    pipeline_id="django_cog_pipeline", step=step or "processing"
-                )
-            elif status == "completed":
-                processed_data.mark_processing_completed()
-            elif status == "failed":
-                processed_data.mark_processing_failed(f"Pipeline step failed: {step}")
-
-        return {
-            "status": "success",
-            "capture_uuid": capture_uuid,
-            "processing_type": processing_type,
-            "status": status,
-        }
-    except Exception as e:
-        logger.error(f"Failed to update processing status: {e}")
-        raise
-
-
-@cog
-def setup_post_processing_cog(
-    capture_uuid: str, processing_types: list[str] | None = None
-) -> dict[str, Any]:
-    """Setup post-processing for a capture.
-
-    Args:
-        capture_uuid: UUID of the capture to process
-        processing_types: List of processing types to run
-
-    Returns:
-        Dict with status and setup info
-    """
-    try:
-        logger.info(f"Starting setup for capture {capture_uuid}")
-
-        # Import models here to avoid Django app registry issues
-        from .models import Capture
-        from .models import CaptureType
-        from .models import ProcessingType
-        from .tasks import _create_or_reset_processed_data
-
-        # Get the capture with retry mechanism for transaction timing issues
-        capture: Capture | None = None
-        max_retries = 3
-        retry_delay = 1  # seconds
-
-        for attempt in range(max_retries):
-            try:
-                capture = Capture.objects.get(uuid=capture_uuid, is_deleted=False)
-                break  # Found the capture, exit retry loop
-            except Capture.DoesNotExist:
-                if attempt < max_retries - 1:
-                    logger.warning(
-                        f"Capture {capture_uuid} not found on attempt {attempt + 1}, "
-                        f"retrying in {retry_delay} seconds..."
-                    )
-                    import time
-
-                    time.sleep(retry_delay)
-                    retry_delay *= 2  # Exponential backoff
-                else:
-                    # Final attempt failed
-                    error_msg = (
-                        f"Capture {capture_uuid} not found after {max_retries} attempts"
-                    )
-                    logger.error(error_msg)
-                    raise ValueError(error_msg)
-
-        # At this point, capture should not be None due to the retry logic above
-        assert capture is not None, (
-            f"Capture {capture_uuid} should have been found by now"
-        )
-
-        # Validate capture type
-        if capture.capture_type != CaptureType.DigitalRF:
-            raise ValueError(f"Capture {capture_uuid} is not a DigitalRF capture")
-
-        # Set default processing types if not specified
-        if not processing_types:
-            processing_types = [ProcessingType.Waterfall.value]
-
-        # Create PostProcessedData records for each processing type
-        for processing_type in processing_types:
-            _create_or_reset_processed_data(capture, processing_type)
-
-        logger.info(f"Completed setup for capture {capture_uuid}")
-        return {
-            "status": "success",
-            "capture_uuid": capture_uuid,
-            "processing_types": processing_types,
-            "capture_type": capture.capture_type.value,
-            "channel": capture.channel,
-        }
-    except Exception as e:
-        logger.error(f"Setup failed for capture {capture_uuid}: {e}")
-        raise
 
 
 # Pipeline configuration functions for Django admin setup
@@ -232,6 +24,7 @@ def get_waterfall_pipeline_config() -> dict[str, Any]:
     Pipeline:
     - Name: "Waterfall Processing"
     - Schedule: None (manual launch)
+    - Prevent overlapping runs: False (allow concurrent processing)
 
     Stages:
     1. "setup_stage" - setup_post_processing_cog (validates capture, creates records)
@@ -243,6 +36,7 @@ def get_waterfall_pipeline_config() -> dict[str, Any]:
     """
     return {
         "pipeline_name": "Waterfall Processing",
+        "prevent_overlapping_runs": False,
         "stages": [
             {
                 "name": "setup_stage",
@@ -293,3 +87,543 @@ def get_pipeline_config(pipeline_type: str) -> dict[str, Any]:
         raise ValueError(f"Unknown pipeline type: {pipeline_type}")
 
     return config_func()
+
+
+# Cog functions (pipeline steps)
+@cog
+def setup_post_processing_cog(
+    capture_uuid: str, processing_types: list[str] | None = None
+) -> dict[str, Any]:
+    """Setup post-processing for a capture.
+
+    Args:
+        capture_uuid: UUID of the capture to process
+        processing_types: List of processing types to run
+
+    Returns:
+        Dict with status and setup info
+    """
+    try:
+        logger.info(f"Starting setup for capture {capture_uuid}")
+
+        # Import models here to avoid Django app registry issues
+        from .models import Capture
+        from .models import CaptureType
+        from .models import ProcessingType
+
+        # Get the capture with retry mechanism for transaction timing issues
+        capture: Capture | None = None
+        max_retries = 3
+        retry_delay = 1  # seconds
+
+        for attempt in range(max_retries):
+            try:
+                capture = Capture.objects.get(uuid=capture_uuid, is_deleted=False)
+                break  # Found the capture, exit retry loop
+            except Capture.DoesNotExist:
+                if attempt < max_retries - 1:
+                    logger.warning(
+                        f"Capture {capture_uuid} not found on attempt {attempt + 1}, "
+                        f"retrying in {retry_delay} seconds..."
+                    )
+                    import time
+
+                    time.sleep(retry_delay)
+                    retry_delay *= 2  # Exponential backoff
+                else:
+                    # Final attempt failed
+                    error_msg = (
+                        f"Capture {capture_uuid} not found after {max_retries} attempts"
+                    )
+                    logger.error(error_msg)
+                    raise ValueError(error_msg)
+
+        # At this point, capture should not be None due to the retry logic above
+        assert capture is not None, (
+            f"Capture {capture_uuid} should have been found by now"
+        )
+
+        # Validate capture type
+        if capture.capture_type != CaptureType.DigitalRF:
+            raise ValueError(f"Capture {capture_uuid} is not a DigitalRF capture")
+
+        # Set default processing types if not specified
+        if not processing_types:
+            processing_types = [ProcessingType.Waterfall.value]
+
+        # Create PostProcessedData records for each processing type
+        for processing_type in processing_types:
+            processed_data = _create_or_reset_processed_data(capture, processing_type)
+            # Mark as ready for processing (but not started yet)
+            processed_data.update_pipeline_step("setup_completed")
+
+        logger.info(f"Completed setup for capture {capture_uuid}")
+        return {
+            "status": "success",
+            "capture_uuid": capture_uuid,
+            "processing_types": processing_types,
+            "capture_type": capture.capture_type,
+            "channel": capture.channel,
+        }
+    except Exception as e:
+        logger.error(f"Setup failed for capture {capture_uuid}: {e}")
+        raise
+
+
+@cog
+def process_waterfall_data_cog(
+    capture_uuid: str, max_slices: int = 100
+) -> dict[str, Any]:
+    """Process waterfall data for a capture.
+
+    Args:
+        capture_uuid: UUID of the capture to process
+        max_slices: Maximum number of slices to process (default: 100)
+
+    Returns:
+        Dict with status and processed data info
+    """
+    logger.info(f"Processing waterfall JSON data for capture {capture_uuid}")
+
+    try:
+        # Import models here to avoid Django app registry issues
+        from .models import Capture
+        from .models import PostProcessedData
+        from .models import ProcessingType
+
+        capture = Capture.objects.get(uuid=capture_uuid, is_deleted=False)
+
+        # Get the processed data record and mark processing as started
+        processed_data = PostProcessedData.objects.filter(
+            capture=capture,
+            processing_type=ProcessingType.Waterfall.value,
+        ).first()
+
+        if not processed_data:
+            return {
+                "status": "error",
+                "message": "No processed data record found for waterfall processing",
+            }
+
+        # Mark processing as started
+        processed_data.mark_processing_started(
+            pipeline_id="django_cog_pipeline", step="waterfall_processing"
+        )
+
+        # Create temporary directory for processing
+        temp_dir = tempfile.mkdtemp(prefix=f"waterfall_{capture_uuid}_")
+        temp_path = Path(temp_dir)
+
+        try:
+            # Update step to file reconstruction
+            processed_data.update_pipeline_step("reconstructing_files")
+
+            # Reconstruct the DigitalRF files for processing
+            capture_files = capture.files.filter(is_deleted=False)
+            reconstructed_path = reconstruct_drf_files(
+                capture, capture_files, temp_path
+            )
+
+            if not reconstructed_path:
+                processed_data.mark_processing_failed(
+                    "Failed to reconstruct DigitalRF directory structure"
+                )
+                return {
+                    "status": "error",
+                    "message": "Failed to reconstruct DigitalRF directory structure",
+                }
+
+            # Update step to data conversion
+            processed_data.update_pipeline_step("converting_data")
+
+            # Process the waterfall data in JSON format
+            waterfall_result = convert_drf_to_waterfall_json(
+                reconstructed_path,
+                capture.channel,
+                ProcessingType.Waterfall.value,
+                max_slices,
+            )
+
+            if waterfall_result["status"] != "success":
+                processed_data.mark_processing_failed(waterfall_result["message"])
+                return waterfall_result
+
+            # Update step to file creation
+            processed_data.update_pipeline_step("storing_data")
+
+            # Create a temporary JSON file
+            import json
+
+            with tempfile.NamedTemporaryFile(
+                mode="w", suffix=".json", delete=False
+            ) as temp_file:
+                json.dump(waterfall_result["json_data"], temp_file, indent=2)
+                temp_file_path = temp_file.name
+
+            # Store the JSON file
+            filename = f"waterfall_{capture_uuid}.json"
+            store_result = store_processed_data(
+                capture_uuid,
+                ProcessingType.Waterfall.value,
+                temp_file_path,
+                filename,
+                waterfall_result["metadata"],
+            )
+
+            # Clean up temporary file
+            os.unlink(temp_file_path)
+
+            if store_result["status"] != "success":
+                processed_data.mark_processing_failed(store_result["message"])
+                return store_result
+
+            # Mark processing as completed
+            processed_data.mark_processing_completed()
+            processed_data.update_pipeline_step("completed")
+
+            logger.info(f"Completed waterfall processing for capture {capture_uuid}")
+            return {
+                "status": "success",
+                "capture_uuid": capture_uuid,
+                "message": "Waterfall JSON data processed and stored successfully",
+                "json_data": waterfall_result["json_data"],
+                "metadata": waterfall_result["metadata"],
+                "store_result": store_result,
+            }
+
+        except Exception as e:
+            # Mark processing as failed
+            processed_data.mark_processing_failed(f"Processing failed: {e!s}")
+            raise
+
+    except Exception as e:
+        logger.error(f"Waterfall processing failed for capture {capture_uuid}: {e}")
+        raise
+
+
+# Helper functions
+def _create_or_reset_processed_data(capture, processing_type: str):
+    """Create or reset a PostProcessedData record for a capture and processing type.
+
+    Args:
+        capture: The capture to create processed data for
+        processing_type: Type of processing (waterfall, spectrogram, etc.)
+
+    Returns:
+        PostProcessedData: The created or reset record
+    """
+    # Import models here to avoid Django app registry issues
+    from sds_gateway.api_methods.models import PostProcessedData
+    from sds_gateway.api_methods.models import ProcessingStatus
+
+    # Try to get existing record
+    processed_data, newly_created = PostProcessedData.objects.get_or_create(
+        capture=capture,
+        processing_type=processing_type,
+        processing_parameters={},  # Default empty parameters
+        defaults={
+            "processing_status": ProcessingStatus.Pending.value,
+            "metadata": {},
+        },
+    )
+
+    if not newly_created:
+        # Reset existing record
+        processed_data.processing_status = ProcessingStatus.Pending.value
+        processed_data.processing_error = ""
+        processed_data.processed_at = None
+        processed_data.pipeline_id = ""
+        processed_data.pipeline_step = ""
+        processed_data.metadata = {}
+        if processed_data.data_file:
+            processed_data.data_file.delete(save=False)
+        processed_data.save()
+
+    return processed_data
+
+
+def reconstruct_drf_files(capture, capture_files, temp_path: Path) -> Path | None:
+    """Reconstruct DigitalRF directory structure from SDS files."""
+    # Import utilities here to avoid Django app registry issues
+    from sds_gateway.api_methods.utils.minio_client import get_minio_client
+
+    logger.info("Reconstructing DigitalRF directory structure")
+
+    try:
+        minio_client = get_minio_client()
+
+        # Create the capture directory structure
+        capture_dir = temp_path / str(capture.uuid)
+        capture_dir.mkdir(parents=True, exist_ok=True)
+
+        # Download and place files in the correct structure
+        for file_obj in capture_files:
+            # Create the directory structure
+            file_path = capture_dir / file_obj.directory.lstrip("/") / file_obj.name
+            file_path.parent.mkdir(parents=True, exist_ok=True)
+
+            # Download the file from MinIO
+            minio_client.fget_object(
+                bucket_name=settings.AWS_STORAGE_BUCKET_NAME,
+                object_name=file_obj.file.name,
+                file_path=str(file_path),
+            )
+
+        # Find the DigitalRF root directory (parent of the channel directory)
+        for root, dirs, files in os.walk(capture_dir):
+            if "drf_properties.h5" in files:
+                # The DigitalRF root is the parent of the channel directory
+                drf_root = Path(root).parent
+                logger.info(f"Found DigitalRF root at: {drf_root}")
+                return drf_root
+
+        logger.error("Could not find DigitalRF properties file")
+        return None
+
+    except Exception as e:
+        logger.exception(f"Error reconstructing DigitalRF files: {e}")
+        return None
+
+
+def convert_drf_to_waterfall_json(
+    drf_path: Path, channel: str, processing_type: str, max_slices: int = 100
+) -> dict:
+    """Convert DigitalRF data to waterfall JSON format similar to SVI implementation."""
+    logger.info(
+        f"Converting DigitalRF data to waterfall JSON format for channel {channel}"
+    )
+
+    try:
+        # Initialize DigitalRF reader
+        reader = DigitalRFReader(str(drf_path))
+        channels = reader.get_channels()
+
+        if not channels:
+            return {
+                "status": "error",
+                "message": "No channels found in DigitalRF data",
+            }
+
+        if channel not in channels:
+            return {
+                "status": "error",
+                "message": f"Channel {channel} not found in DigitalRF data. Available channels: {channels}",
+            }
+
+        # Get sample bounds
+        bounds = reader.get_bounds(channel)
+        if bounds is None:
+            return {
+                "status": "error",
+                "message": "Could not get sample bounds for channel",
+            }
+
+        start_sample, end_sample = bounds
+        if start_sample is None or end_sample is None:
+            return {
+                "status": "error",
+                "message": "Invalid sample bounds for channel",
+            }
+        total_samples = end_sample - start_sample
+
+        # Get metadata from DigitalRF properties
+        drf_props_path = drf_path / channel / "drf_properties.h5"
+        with h5py.File(drf_props_path, "r") as f:
+            sample_rate_numerator = f.attrs.get("sample_rate_numerator")
+            sample_rate_denominator = f.attrs.get("sample_rate_denominator")
+            if sample_rate_numerator is None or sample_rate_denominator is None:
+                return {
+                    "status": "error",
+                    "message": "Sample rate information missing from DigitalRF properties",
+                }
+            sample_rate = float(sample_rate_numerator) / float(sample_rate_denominator)
+
+        # Get center frequency from metadata
+        center_freq = 0.0
+        try:
+            # Try to get center frequency from metadata
+            metadata_dict = reader.read_metadata(
+                start_sample, min(1000, end_sample - start_sample), channel
+            )
+            if metadata_dict and "center_freq" in metadata_dict:
+                center_freq = float(metadata_dict["center_freq"])
+        except Exception as e:
+            logger.warning(f"Could not read center frequency from metadata: {e}")
+
+        # Calculate frequency range
+        freq_span = sample_rate
+        min_frequency = center_freq - freq_span / 2
+        max_frequency = center_freq + freq_span / 2
+
+        # Processing parameters
+        fft_size = 1024  # Default, could be configurable
+        samples_per_slice = 1024  # Default, could be configurable
+
+        # Calculate total slices and limit to max_slices
+        total_slices = total_samples // samples_per_slice
+        slices_to_process = min(total_slices, max_slices)
+
+        logger.info(
+            f"Processing {slices_to_process} slices with {samples_per_slice} samples per slice"
+        )
+
+        # Process slices and create JSON data
+        waterfall_data = []
+
+        for slice_idx in range(slices_to_process):
+            # Calculate sample range for this slice
+            slice_start_sample = start_sample + slice_idx * samples_per_slice
+            slice_num_samples = min(samples_per_slice, end_sample - slice_start_sample)
+
+            if slice_num_samples <= 0:
+                break
+
+            # Read the data
+            data_array = reader.read_vector(
+                slice_start_sample, slice_num_samples, channel, 0
+            )
+
+            # Perform FFT processing
+            fft_data = np.fft.fft(data_array, n=fft_size)
+            power_spectrum = np.abs(fft_data) ** 2
+
+            # Convert to dB
+            power_spectrum_db = 10 * np.log10(power_spectrum + 1e-12)
+
+            # Convert power spectrum to binary string for transmission
+            data_bytes = power_spectrum_db.astype(np.float32).tobytes()
+            data_string = base64.b64encode(data_bytes).decode("utf-8")
+
+            # Create timestamp from sample index
+            timestamp = datetime.datetime.fromtimestamp(
+                slice_start_sample / sample_rate, tz=datetime.UTC
+            ).isoformat()
+
+            # Build WaterfallFile format with enhanced metadata
+            waterfall_file = {
+                "data": data_string,
+                "data_type": "float32",
+                "timestamp": timestamp,
+                "min_frequency": min_frequency,
+                "max_frequency": max_frequency,
+                "num_samples": slice_num_samples,
+                "sample_rate": sample_rate,
+                "center_frequency": center_freq,
+            }
+
+            # Build custom fields with all available metadata
+            custom_fields = {
+                "channel_name": channel,
+                "start_sample": slice_start_sample,
+                "num_samples": slice_num_samples,
+                "fft_size": fft_size,
+                "scan_time": slice_num_samples / sample_rate,
+                "slice_index": slice_idx,
+            }
+
+            waterfall_file["custom_fields"] = custom_fields
+            waterfall_data.append(waterfall_file)
+
+            # Log progress every 10 slices
+            if slice_idx % 10 == 0:
+                logger.info(f"Processed {slice_idx}/{slices_to_process} slices")
+
+        metadata = {
+            "center_frequency": center_freq,
+            "sample_rate": sample_rate,
+            "min_frequency": min_frequency,
+            "max_frequency": max_frequency,
+            "total_slices": total_slices,
+            "slices_processed": len(waterfall_data),
+            "fft_size": fft_size,
+            "samples_per_slice": samples_per_slice,
+            "channel": channel,
+        }
+
+        return {
+            "status": "success",
+            "message": "Waterfall data converted to JSON successfully",
+            "json_data": waterfall_data,
+            "metadata": metadata,
+        }
+
+    except Exception as e:
+        logger.exception(f"Error converting DigitalRF to waterfall JSON: {e}")
+        return {
+            "status": "error",
+            "message": f"Error converting DigitalRF data: {e}",
+        }
+
+
+def store_processed_data(
+    capture_uuid: str,
+    processing_type: str,
+    file_path: str,
+    filename: str,
+    metadata: dict | None = None,
+) -> dict:
+    """
+    Store processed data back to SDS storage as a file.
+
+    Args:
+        capture_uuid: UUID of the capture
+        processing_type: Type of processed data (waterfall, spectrogram, etc.)
+        file_path: Path to the file to store
+        filename: Name for the stored file
+        metadata: Metadata to store
+
+    Returns:
+        dict: Storage result
+    """
+    # Import models here to avoid Django app registry issues
+    from sds_gateway.api_methods.models import Capture
+    from sds_gateway.api_methods.models import PostProcessedData
+
+    logger.info(f"Storing {processing_type} file for capture {capture_uuid}")
+
+    try:
+        capture = Capture.objects.get(uuid=capture_uuid, is_deleted=False)
+
+        # Get the processed data record
+        processed_data = PostProcessedData.objects.filter(
+            capture=capture,
+            processing_type=processing_type,
+        ).first()
+
+        if not processed_data:
+            return {
+                "status": "error",
+                "message": f"No processed data record found for {processing_type}",
+            }
+
+        # Store the file
+        processed_data.set_processed_data_file(file_path, filename)
+
+        # Update metadata if provided
+        if metadata:
+            processed_data.metadata.update(metadata)
+
+        # Add storage metadata
+        processed_data.metadata.update(
+            {
+                "stored_at": datetime.datetime.now(datetime.UTC).isoformat(),
+                "data_format": "file",
+                "file_size": processed_data.data_file.size,
+            }
+        )
+        processed_data.save()
+
+        logger.info(f"Stored file {filename} for {processing_type} data")
+
+        return {
+            "status": "success",
+            "message": f"{processing_type} file stored successfully",
+            "file_name": filename,
+        }
+
+    except Exception as e:
+        logger.exception(f"Error storing {processing_type} file: {e}")
+        return {
+            "status": "error",
+            "message": f"Error storing {processing_type} file: {e}",
+        }
