+++ gateway/sds_gateway/api_methods/tasks.py
@@ -1,5 +1,4 @@
 import datetime
-import os
 import re
 import uuid
 import zipfile
@@ -19,7 +18,6 @@ from sds_gateway.api_methods.models import Capture
 from sds_gateway.api_methods.models import Dataset
 from sds_gateway.api_methods.models import File
 from sds_gateway.api_methods.models import ItemType
-from sds_gateway.api_methods.models import PostProcessedData
 from sds_gateway.api_methods.models import ProcessingType
 from sds_gateway.api_methods.models import TemporaryZipFile
 from sds_gateway.api_methods.models import user_has_access_to_item
@@ -1057,291 +1055,4 @@ def start_capture_post_processing(
             "status": "error",
             "message": error_msg,
             "capture_uuid": capture_uuid,
-        }
-
-
-@shared_task
-def store_processed_data(
-    capture_uuid: str,
-    processing_type: str,
-    file_path: str,
-    filename: str,
-    metadata: dict | None = None,
-) -> dict:
-    """
-    Store processed data back to SDS storage as a file.
-
-    Args:
-        capture_uuid: UUID of the capture
-        processing_type: Type of processed data (waterfall, spectrogram, etc.)
-        file_path: Path to the file to store
-        filename: Name for the stored file
-        metadata: Metadata to store
-
-    Returns:
-        dict: Storage result
-    """
-    logger.info(f"Storing {processing_type} file for capture {capture_uuid}")
-
-    try:
-        capture = Capture.objects.get(uuid=capture_uuid, is_deleted=False)
-
-        # Get the processed data record
-        processed_data = PostProcessedData.objects.filter(
-            capture=capture,
-            processing_type=processing_type,
-        ).first()
-
-        if not processed_data:
-            return {
-                "status": "error",
-                "message": f"No processed data record found for {processing_type}",
-            }
-
-        # Store the file
-        processed_data.set_processed_data_file(file_path, filename)
-
-        # Update metadata if provided
-        if metadata:
-            processed_data.metadata.update(metadata)
-
-        # Add storage metadata
-        processed_data.metadata.update(
-            {
-                "stored_at": datetime.datetime.now(datetime.UTC).isoformat(),
-                "data_format": "file",
-                "file_size": processed_data.data_file.size,
-            }
-        )
-        processed_data.save()
-
-        logger.info(f"Stored file {filename} for {processing_type} data")
-
-        return {
-            "status": "success",
-            "message": f"{processing_type} file stored successfully",
-            "file_name": filename,
-        }
-
-    except Exception as e:
-        logger.exception(f"Error storing {processing_type} file: {e}")
-        return {
-            "status": "error",
-            "message": f"Error storing {processing_type} file: {e}",
-        }
-
-
-# Helper functions (existing implementations)
-def reconstruct_drf_files(
-    capture: Capture, capture_files, temp_path: Path
-) -> Path | None:
-    """Reconstruct DigitalRF directory structure from SDS files."""
-    from sds_gateway.api_methods.utils.minio_client import get_minio_client
-
-    logger.info("Reconstructing DigitalRF directory structure")
-
-    try:
-        minio_client = get_minio_client()
-
-        # Create the capture directory structure
-        capture_dir = temp_path / str(capture.uuid)
-        capture_dir.mkdir(parents=True, exist_ok=True)
-
-        # Download and place files in the correct structure
-        for file_obj in capture_files:
-            # Create the directory structure
-            file_path = capture_dir / file_obj.directory.lstrip("/") / file_obj.name
-            file_path.parent.mkdir(parents=True, exist_ok=True)
-
-            # Download the file from MinIO
-            minio_client.fget_object(
-                bucket_name=settings.AWS_STORAGE_BUCKET_NAME,
-                object_name=file_obj.file.name,
-                file_path=str(file_path),
-            )
-
-        # Find the DigitalRF root directory (parent of the channel directory)
-        for root, dirs, files in os.walk(capture_dir):
-            if "drf_properties.h5" in files:
-                # The DigitalRF root is the parent of the channel directory
-                drf_root = Path(root).parent
-                logger.info(f"Found DigitalRF root at: {drf_root}")
-                return drf_root
-
-        logger.error("Could not find DigitalRF properties file")
-        return None
-
-    except Exception as e:
-        logger.exception(f"Error reconstructing DigitalRF files: {e}")
-        return None
-
-
-def convert_drf_to_waterfall_json(
-    drf_path: Path, channel: str, processing_type: str, max_slices: int = 100
-) -> dict:
-    """Convert DigitalRF data to waterfall JSON format similar to SVI implementation."""
-    import base64
-
-    import h5py
-    import numpy as np
-    from digital_rf import DigitalRFReader
-
-    logger.info(
-        f"Converting DigitalRF data to waterfall JSON format for channel {channel}"
-    )
-
-    try:
-        # Initialize DigitalRF reader
-        reader = DigitalRFReader(str(drf_path))
-        channels = reader.get_channels()
-
-        if not channels:
-            return {
-                "status": "error",
-                "message": "No channels found in DigitalRF data",
-            }
-
-        if channel not in channels:
-            return {
-                "status": "error",
-                "message": f"Channel {channel} not found in DigitalRF data. Available channels: {channels}",
-            }
-
-        # Get sample bounds
-        bounds = reader.get_bounds(channel)
-        if bounds is None:
-            return {
-                "status": "error",
-                "message": "Could not get sample bounds for channel",
-            }
-
-        start_sample, end_sample = bounds
-        total_samples = end_sample - start_sample
-
-        # Get metadata from DigitalRF properties
-        drf_props_path = drf_path / channel / "drf_properties.h5"
-        with h5py.File(drf_props_path, "r") as f:
-            sample_rate_numerator = f.attrs.get("sample_rate_numerator")
-            sample_rate_denominator = f.attrs.get("sample_rate_denominator")
-            if sample_rate_numerator is None or sample_rate_denominator is None:
-                return {
-                    "status": "error",
-                    "message": "Sample rate information missing from DigitalRF properties",
-                }
-            sample_rate = float(sample_rate_numerator) / float(sample_rate_denominator)
-
-        # Get center frequency from metadata
-        center_freq = 0.0
-        try:
-            # Try to get center frequency from metadata
-            metadata_dict = reader.read_metadata(
-                start_sample, min(1000, end_sample - start_sample), channel
-            )
-            if metadata_dict and "center_freq" in metadata_dict:
-                center_freq = float(metadata_dict["center_freq"])
-        except Exception as e:
-            logger.warning(f"Could not read center frequency from metadata: {e}")
-
-        # Calculate frequency range
-        freq_span = sample_rate
-        min_frequency = center_freq - freq_span / 2
-        max_frequency = center_freq + freq_span / 2
-
-        # Processing parameters
-        fft_size = 1024  # Default, could be configurable
-        samples_per_slice = 1024  # Default, could be configurable
-
-        # Calculate total slices and limit to max_slices
-        total_slices = total_samples // samples_per_slice
-        slices_to_process = min(total_slices, max_slices)
-
-        logger.info(
-            f"Processing {slices_to_process} slices with {samples_per_slice} samples per slice"
-        )
-
-        # Process slices and create JSON data
-        waterfall_data = []
-
-        for slice_idx in range(slices_to_process):
-            # Calculate sample range for this slice
-            slice_start_sample = start_sample + slice_idx * samples_per_slice
-            slice_num_samples = min(samples_per_slice, end_sample - slice_start_sample)
-
-            if slice_num_samples <= 0:
-                break
-
-            # Read the data
-            data_array = reader.read_vector(
-                slice_start_sample, slice_num_samples, channel, 0
-            )
-
-            # Perform FFT processing
-            fft_data = np.fft.fft(data_array, n=fft_size)
-            power_spectrum = np.abs(fft_data) ** 2
-
-            # Convert to dB
-            power_spectrum_db = 10 * np.log10(power_spectrum + 1e-12)
-
-            # Convert power spectrum to binary string for transmission
-            data_bytes = power_spectrum_db.astype(np.float32).tobytes()
-            data_string = base64.b64encode(data_bytes).decode("utf-8")
-
-            # Create timestamp from sample index
-            timestamp = datetime.datetime.fromtimestamp(
-                slice_start_sample / sample_rate, tz=datetime.UTC
-            ).isoformat()
-
-            # Build WaterfallFile format with enhanced metadata
-            waterfall_file = {
-                "data": data_string,
-                "data_type": "float32",
-                "timestamp": timestamp,
-                "min_frequency": min_frequency,
-                "max_frequency": max_frequency,
-                "num_samples": slice_num_samples,
-                "sample_rate": sample_rate,
-                "center_frequency": center_freq,
-            }
-
-            # Build custom fields with all available metadata
-            custom_fields = {
-                "channel_name": channel,
-                "start_sample": slice_start_sample,
-                "num_samples": slice_num_samples,
-                "fft_size": fft_size,
-                "scan_time": slice_num_samples / sample_rate,
-                "slice_index": slice_idx,
-            }
-
-            waterfall_file["custom_fields"] = custom_fields
-            waterfall_data.append(waterfall_file)
-
-            # Log progress every 10 slices
-            if slice_idx % 10 == 0:
-                logger.info(f"Processed {slice_idx}/{slices_to_process} slices")
-
-        metadata = {
-            "center_frequency": center_freq,
-            "sample_rate": sample_rate,
-            "min_frequency": min_frequency,
-            "max_frequency": max_frequency,
-            "total_slices": total_slices,
-            "slices_processed": len(waterfall_data),
-            "fft_size": fft_size,
-            "samples_per_slice": samples_per_slice,
-            "channel": channel,
-        }
-
-        return {
-            "status": "success",
-            "message": "Waterfall data converted to JSON successfully",
-            "json_data": waterfall_data,
-            "metadata": metadata,
-        }
-
-    except Exception as e:
-        logger.exception(f"Error converting DigitalRF to waterfall JSON: {e}")
-        return {
-            "status": "error",
-            "message": f"Error converting DigitalRF data: {e}",
         }
